+++ apps/decodex/src/agent/app_server/server_requests.rs
@@ -1,336 +1,15 @@
 //! App-server server request routing and non-interactive response handling.
 
-use color_eyre::eyre::Report;
+mod dispatch;
+mod recording;
+mod rejection;
 
-use crate::{
-	agent::app_server::{
-		self, AppServerClient, ChatgptAuthTokensRefreshParams, ChatgptAuthTokensRefreshResponse,
-		CommandExecutionApprovalDecision, CommandExecutionRequestApprovalResponse,
-		FileChangeApprovalDecision, FileChangeRequestApprovalResponse, JSONRPC_METHOD_NOT_FOUND,
-		JsonRpcConnection, JsonRpcMessage, JsonRpcRequest, McpServerElicitationAction,
-		McpServerElicitationRequestResponse, PermissionGrantScope,
-		PermissionsRequestApprovalResponse, RequestDispatchContext, RequestWaitPhase, RunRecorder,
-		Serialize, ThreadStatusChangedNotification, ToolRequestUserInputResponse, WireMessage,
-		eyre, redact_identifier, serde_json,
+#[cfg(test)]
+pub(super) use self::recording::{record_interactive_request_state, record_server_request};
+pub(super) use self::{
+	dispatch::{handle_server_request_during_turn_execution, handle_server_request_while_waiting},
+	recording::{
+		apply_protocol_message_side_effects, interactive_flag_for_request,
+		record_server_request_response,
 	},
-	prelude::Result,
 };
-
-pub(super) fn handle_server_request_while_waiting(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	wire_message: &WireMessage,
-	request: &JsonRpcRequest,
-	context: RequestDispatchContext<'_>,
-) -> Result<()> {
-	if app_server::targets_thread(wire_message, context.target_thread_id) {
-		record_wire_message_safely(recorder, wire_message)?;
-		record_interactive_request_state(recorder, request)?;
-	} else if request.method == "account/chatgptAuthTokens/refresh" {
-		record_codex_account_refresh_request(recorder, request)?;
-	}
-
-	dispatch_server_request(connection, recorder, request, context)
-}
-
-pub(super) fn handle_server_request_during_turn_execution(
-	client: &mut AppServerClient,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	context: RequestDispatchContext<'_>,
-) -> Result<()> {
-	record_server_request_safely(recorder, request)?;
-	record_interactive_request_state(recorder, request)?;
-
-	dispatch_server_request(&mut client.connection, recorder, request, context)
-}
-
-pub(super) fn record_server_request(
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-) -> Result<()> {
-	recorder.record(
-		request.method.as_str(),
-		&serde_json::json!({
-			"id": request.id.clone(),
-			"method": request.method.clone(),
-			"params": request.params.clone(),
-		})
-		.to_string(),
-	)
-}
-
-pub(super) fn record_server_request_response<T>(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	event_type: &str,
-	response: &T,
-) -> Result<()>
-where
-	T: Serialize,
-{
-	connection.respond(&request.id, response)?;
-
-	recorder.record(event_type, &serde_json::to_string(response)?)
-}
-
-pub(super) fn record_interactive_request_state(
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-) -> Result<()> {
-	let Some(flag) = interactive_flag_for_request(request.method.as_str()) else {
-		return Ok(());
-	};
-
-	if let Some(thread_id) = app_server::thread_id_from_value(&request.params) {
-		recorder.set_thread_id(thread_id)?;
-	}
-	if let Some(turn_id) = app_server::turn_id_from_value(&request.params) {
-		recorder.set_turn_id(turn_id)?;
-	}
-
-	recorder.set_thread_status("active", &[flag.to_owned()])
-}
-
-pub(super) fn interactive_flag_for_request(method: &str) -> Option<&'static str> {
-	match method {
-		"item/tool/requestUserInput" => Some("waitingOnUserInput"),
-		"item/commandExecution/requestApproval"
-		| "item/fileChange/requestApproval"
-		| "item/permissions/requestApproval"
-		| "mcpServer/elicitation/request" => Some("waitingOnApproval"),
-		_ => None,
-	}
-}
-
-pub(super) fn apply_protocol_message_side_effects(
-	recorder: &mut RunRecorder<'_>,
-	message: &WireMessage,
-) -> Result<()> {
-	match &message.message {
-		JsonRpcMessage::Notification(notification)
-			if notification.method == "thread/status/changed" =>
-		{
-			let payload: ThreadStatusChangedNotification =
-				serde_json::from_value(notification.params.clone())?;
-
-			if recorder.thread_id.is_none() {
-				recorder.set_thread_id(&payload.thread_id)?;
-			}
-
-			recorder.set_thread_status(&payload.status.kind, &payload.status.active_flags)?;
-		},
-		_ => {},
-	}
-
-	Ok(())
-}
-
-fn dispatch_server_request(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	context: RequestDispatchContext<'_>,
-) -> Result<()> {
-	match request.method.as_str() {
-		"item/tool/call" if context.phase == RequestWaitPhase::TurnExecution =>
-			app_server::dispatch_dynamic_tool_call(connection, recorder, request, context),
-		"account/chatgptAuthTokens/refresh" =>
-			dispatch_codex_account_refresh(connection, recorder, request, context),
-		"item/tool/call" => app_server::respond_to_dynamic_tool_call_dispatch(
-			connection,
-			recorder,
-			request,
-			app_server::dynamic_tool_call_unavailable_for_phase(context.phase),
-		),
-		"item/commandExecution/requestApproval" => reject_interactive_server_request(
-			connection,
-			recorder,
-			request,
-			context.phase,
-			"item/commandExecution/requestApproval/response",
-			&CommandExecutionRequestApprovalResponse {
-				decision: CommandExecutionApprovalDecision::Decline,
-			},
-		),
-		"item/fileChange/requestApproval" => reject_interactive_server_request(
-			connection,
-			recorder,
-			request,
-			context.phase,
-			"item/fileChange/requestApproval/response",
-			&FileChangeRequestApprovalResponse { decision: FileChangeApprovalDecision::Decline },
-		),
-		"item/tool/requestUserInput" => reject_interactive_server_request(
-			connection,
-			recorder,
-			request,
-			context.phase,
-			"item/tool/requestUserInput/response",
-			&ToolRequestUserInputResponse::default(),
-		),
-		"item/permissions/requestApproval" => reject_interactive_server_request(
-			connection,
-			recorder,
-			request,
-			context.phase,
-			"item/permissions/requestApproval/response",
-			&PermissionsRequestApprovalResponse {
-				permissions: Default::default(),
-				scope: PermissionGrantScope::Turn,
-			},
-		),
-		"mcpServer/elicitation/request" => reject_interactive_server_request(
-			connection,
-			recorder,
-			request,
-			context.phase,
-			"mcpServer/elicitation/request/response",
-			&McpServerElicitationRequestResponse {
-				action: McpServerElicitationAction::Decline,
-				content: None,
-				meta: None,
-			},
-		),
-		other =>
-			reject_unsupported_server_request(connection, recorder, request, context.phase, other),
-	}
-}
-
-fn record_server_request_safely(
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-) -> Result<()> {
-	if request.method == "account/chatgptAuthTokens/refresh" {
-		return record_codex_account_refresh_request(recorder, request);
-	}
-
-	record_server_request(recorder, request)
-}
-
-fn record_wire_message_safely(
-	recorder: &mut RunRecorder<'_>,
-	wire_message: &WireMessage,
-) -> Result<()> {
-	match &wire_message.message {
-		JsonRpcMessage::Request(request)
-			if request.method == "account/chatgptAuthTokens/refresh" =>
-			record_codex_account_refresh_request(recorder, request),
-		_ => recorder.record(app_server::message_type(wire_message), &wire_message.raw),
-	}
-}
-
-fn record_codex_account_refresh_request(
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-) -> Result<()> {
-	let params = serde_json::from_value::<ChatgptAuthTokensRefreshParams>(request.params.clone())
-		.unwrap_or(ChatgptAuthTokensRefreshParams { reason: None, previous_account_id: None });
-
-	recorder.record(
-		"account/chatgptAuthTokens/refresh",
-		&serde_json::json!({
-			"id": request.id.clone(),
-			"method": request.method.as_str(),
-			"reason": params.reason.as_deref(),
-			"previousAccountFingerprint": params.previous_account_id.as_deref().map(redact_identifier),
-		})
-		.to_string(),
-	)
-}
-
-fn dispatch_codex_account_refresh(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	context: RequestDispatchContext<'_>,
-) -> Result<()> {
-	let account_provider = context.codex_account_provider.ok_or_else(|| {
-		eyre::eyre!(
-			"app_server_protocol_failure: received `account/chatgptAuthTokens/refresh` without a configured Codex account provider."
-		)
-	})?;
-	let params = serde_json::from_value::<ChatgptAuthTokensRefreshParams>(request.params.clone())?;
-	let account = match account_provider.refresh_account(params.previous_account_id.as_deref()) {
-		Ok(account) => account,
-		Err(error) => {
-			app_server::record_codex_account_failure(
-				recorder,
-				"account/chatgptAuthTokens/refresh/failed",
-				&error,
-			);
-
-			return Err(error);
-		},
-	};
-	let response = ChatgptAuthTokensRefreshResponse {
-		access_token: account.access_token().to_owned(),
-		chatgpt_account_id: account.account_id().to_owned(),
-		chatgpt_plan_type: account.plan_type().map(str::to_owned),
-	};
-
-	recorder.set_codex_account(account.summary(), account.account_summaries())?;
-	connection.respond(&request.id, &response)?;
-
-	recorder.record(
-		"account/chatgptAuthTokens/refresh/response",
-		&serde_json::json!({
-			"type": "chatgptAuthTokens",
-			"accountFingerprint": account.summary().account_fingerprint.as_str(),
-			"planType": account.summary().plan_type.as_deref(),
-			"refreshStatus": account.summary().refresh_status.as_str(),
-			"primaryRemainingPercent": account.summary().primary_remaining_percent,
-			"secondaryRemainingPercent": account.summary().secondary_remaining_percent,
-		})
-		.to_string(),
-	)
-}
-
-fn reject_unsupported_server_request(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	phase: RequestWaitPhase,
-	method: &str,
-) -> Result<()> {
-	let message = format!("unsupported non-interactive server request `{method}`");
-
-	connection.respond_error(&request.id, JSONRPC_METHOD_NOT_FOUND, &message)?;
-	recorder.record(
-		"json-rpc/error/response",
-		&serde_json::json!({
-			"code": JSONRPC_METHOD_NOT_FOUND,
-			"message": message,
-		})
-		.to_string(),
-	)?;
-
-	eyre::bail!(
-		"app_server_protocol_failure: unsupported server request `{method}` while waiting for {}.",
-		phase.label()
-	);
-}
-
-fn reject_interactive_server_request<T>(
-	connection: &mut JsonRpcConnection,
-	recorder: &mut RunRecorder<'_>,
-	request: &JsonRpcRequest,
-	phase: RequestWaitPhase,
-	event_type: &str,
-	response: &T,
-) -> Result<()>
-where
-	T: Serialize,
-{
-	record_server_request_response(connection, recorder, request, event_type, response)?;
-
-	Err(noninteractive_interaction_required(request.method.as_str(), phase))
-}
-
-fn noninteractive_interaction_required(method: &str, phase: RequestWaitPhase) -> Report {
-	eyre::eyre!(
-		"noninteractive_interaction_required: server request `{method}` requires interactive handling during {}.",
-		phase.label()
-	)
-}
